🎨: 보너스 포인트 계산 및 렌더링 로직을 함수로 추출
